Intermediate: implemented remaining states
The states "started" and "restarted" are now implemented. So
the initial list of states is now complete. Tests are still missing.

As a drive-by fix this also enables the presence of more than one
copy and running instance of Locust.IO by using pkill and pgrep
more selectively when managing processes.