Fixed Locust.IO start command
The command to start Locust.IO was incorrect and returned with an
error. This one does not, but I couldn't test it yet because the
default docker container molecule provides has no init process, so
whatever I start is instantly killed as soon as it detaches from
the Ansible process.