Switched to systemd service and instance_name
The Locust.io slaves are now started as a systemd services.
It is still possible to have several of them installed in
parallel, on the same host, by giving the instances different
instance_name parameters.

This also changed the directory layout. The base directory
is now fixed ('/opt/locust.io') and instances will appear
as subdirectories inside of that.