Switched testing back to docker
After switching molecule to Vagrant/VirtualBox because I didn't
want to run privileged docker containers (and requiring systemd)
I had lots of trouble with dropped (virtual) ethernet connections
in VirtualBox. The CentOS box became unreachable every couple of
minutes and was generally too unreliable to work with.

So now I switched back to docker, made the container privileged and
used an upstream image that has systemd built in. It works, but
I'm not happy about the privileges.